Conflicts:
	intern/cycles/blender/addon/ui.py
	release/datafiles/blender_icons_update.py
	release/datafiles/prvicons.png
	source/blender/blenlib/intern/math_base_inline.c
	source/blender/editors/space_view3d/view3d_edit.c
	source/blender/editors/space_view3d/view3d_intern.h

+++ source/blender/blenlib/intern/math_base_inline.c
@@ -399,8 +399,7 @@ MINLINE int integer_digits_d(const double d)
 
 MINLINE int integer_digits_i(const int i)
 {
-	/*return (int)log10(i) + 1;*/
-	return (int)log10((float)i) + 1; /*bfa-standard defined for both float and int, and so we need to define one of it*/
+	return (int)log10((double)i) + 1;
 }
 
 /* Internal helpers for SSE2 implementation.
